Fall Detector (GPS,SMS alert, Toggle Button to cancel)
The system works by detecting a fall, alerting the user that they have 10 seconds to cancel, through LEDS and a piezo buzzer, by pressing the toggle button. If the button is not pressed the code compiles GPS location and sends it via SMS to a designated number and activates LED and buzzer again.